refactor trade and multiply/divide by precision
