+++ client/apps/game/src/ui/components/hints/resources.tsx
@@ -1,13 +1,14 @@
 import { Headline } from "@/ui/elements/headline";
 import { ResourceCost } from "@/ui/elements/resource-cost";
 import { ResourceIcon } from "@/ui/elements/resource-icon";
-import { currencyFormat, gramToKg, multiplyByPrecision } from "@/ui/utils/utils";
+import { currencyFormat, gramToKg } from "@/ui/utils/utils";
 import {
   CapacityConfigCategory,
   RESOURCE_PRECISION,
   ResourcesIds,
   configManager,
   findResourceById,
+  multiplyByPrecision,
 } from "@bibliothecadao/eternum";
 import { useMemo } from "react";
 import { tableOfContents } from "./utils";
