+++ client/apps/game/src/ui/components/hyperstructures/hyperstructure-resource-chip.tsx
@@ -2,9 +2,9 @@ import { useUIStore } from "@/hooks/store/use-ui-store";
 import Button from "@/ui/elements/button";
 import { NumberInput } from "@/ui/elements/number-input";
 import { ResourceIcon } from "@/ui/elements/resource-icon";
-import { currencyIntlFormat, divideByPrecision } from "@/ui/utils/utils";
+import { currencyIntlFormat } from "@/ui/utils/utils";
 import { getBlockTimestamp } from "@/utils/timestamp";
-import { findResourceById, getBalance, getIconResourceId, ID } from "@bibliothecadao/eternum";
+import { divideByPrecision, findResourceById, getBalance, getIconResourceId, ID } from "@bibliothecadao/eternum";
 import { ProgressWithPercentage, useDojo } from "@bibliothecadao/react";
 import { useEffect, useState } from "react";
