+++ client/apps/game/src/ui/components/worldmap/battles/battle-label.tsx
@@ -2,9 +2,14 @@ import { useUIStore } from "@/hooks/store/use-ui-store";
 import { BaseThreeTooltip, Position } from "@/ui/elements/base-three-tooltip";
 import { Headline } from "@/ui/elements/headline";
 import { DurationLeft, ProgressBar } from "@/ui/modules/military/battle-view/battle-progress";
-import { divideByPrecision } from "@/ui/utils/utils";
 import { getBlockTimestamp } from "@/utils/timestamp";
-import { BattleManager, ContractAddress, getStructureAtPosition, Structure } from "@bibliothecadao/eternum";
+import {
+  BattleManager,
+  ContractAddress,
+  divideByPrecision,
+  getStructureAtPosition,
+  Structure,
+} from "@bibliothecadao/eternum";
 import { useBattlesAtPosition, useDojo, useQuery } from "@bibliothecadao/react";
 import { useMemo } from "react";
