+++ client/apps/landing/src/lib/utils.ts
@@ -1,4 +1,3 @@
-import { RESOURCE_PRECISION } from "@bibliothecadao/eternum";
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 
@@ -17,12 +16,4 @@ export const currencyFormat = (num: number, decimals: number): string => {
 export function displayAddress(string: string) {
   if (string === undefined) return "unknown";
   return string.substring(0, 6) + "..." + string.substring(string.length - 4);
-}
-
-export function multiplyByPrecision(value: number): number {
-  return Math.floor(value * RESOURCE_PRECISION);
-}
-
-export function divideByPrecision(value: number): number {
-  return value / RESOURCE_PRECISION;
 }
